% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_time_to_event_variables.R
\name{add_tte_outcomes}
\alias{add_tte_outcomes}
\title{Add appropriate TTE outcome variables to a data frame based on the df attributes, if the file was
loaded using a load function.}
\usage{
add_tte_outcomes(df)
}
\description{
Uses the helper functions to create three columns for an outcome defined by an event date and a censor date:
\itemize{
\item {prefix}_CENSOR: event date if present, otherwise censor date
\item {prefix}_BINARY: 1 if event date present, else 0
\item {prefix}_{unit}: numeric time from start_date to {prefix}_CENSOR in chosen units
}
}
