% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_strobe_log.R
\name{get_strobe_log}
\alias{get_strobe_log}
\title{Retrieve the STROBE Derivation Log}
\usage{
get_strobe_log()
}
\value{
A data frame with the columns: \code{id}, \code{parent}, \code{label}, \code{filter}, \code{remaining}, \code{dropped}.
}
\description{
Returns the current STROBE tracking data frame containing the sequence of
filtering steps, labels, and counts of included and excluded observations.
}
\examples{
get_strobe_log()
}
